+++ hosted/gke/p1/p1_import_test.go
@@ -29,7 +29,7 @@ var _ = Describe("P1Importing", func() {
 		})
 
 		AfterEach(func() {
-			if ctx.ClusterCleanup {
+			if ctx.ClusterCleanup && cluster != nil {
 				err := helper.DeleteGKEHostCluster(cluster, ctx.RancherAdminClient)
 				Expect(err).To(BeNil())
 				err = helper.DeleteGKEClusterOnGCloud(zone, project, clusterName)
